Fix to RDY pin
RDY needs to be pulled up with a resistor and not linked to VDD directly. Unlike the 65c02, RDY on the 65c816 is bidirectional and can, under some circumstances, attempt to pull itself down. If it's directly connected to VDD, it will attempt to sink VDD and short out. WDC seems to recommend using 3.3K resistors as pullups. I'm not sure how other values would perform. BE is now pulled up with a resistor too, which is recommended by WDC for reasons I'm not sure about.